requirements.txt
Flask
Gunicorn
groq
firebase-admin
python-dotenv